cleanup and move utils to cli-kit
