+++ packages/withstudiocms_cli-kit/package.json
@@ -59,7 +59,8 @@
 		"cli-cursor": "^5.0.0",
 		"is-unicode-supported": "^2.1.0",
 		"slice-ansi": "^7.1.0",
-		"tinyexec": "^1.0.1"
+		"tinyexec": "^1.0.1",
+		"package-manager-detector": "^1.1.0"
 	},
 	"devDependencies": {
 		"@types/figlet": "^1.7.0"
